+++ modyn/selector/internal/selector_strategies/presampling_strategies/trigger_balanced_presampling_strategy.py
@@ -11,13 +11,8 @@ class TriggerBalancedPresamplingStrategy(AbstractBalancedPresamplingStrategy):
 
         self.balanced_column = SelectorStateMetadata.seen_in_trigger_id
 
-    def get_presampling_query(
-        self,
-        next_trigger_id: int,
-        tail_triggers: Optional[int],
-        limit: Optional[int],
-        trigger_dataset_size: Optional[int],
-    ) -> Select:
+    def get_presampling_query(self, next_trigger_id: int, tail_triggers: Optional[int], limit: Optional[int],
+                              trigger_dataset_size: Optional[int], requires_samples_ordered_by_label) -> Select:
         if tail_triggers == 0:
             raise ValueError("You cannot balance across triggers if you use reset_after_trigger")
-        return super().get_presampling_query(next_trigger_id, tail_triggers, limit, trigger_dataset_size)
+        return super().get_presampling_query(next_trigger_id, tail_triggers, limit, trigger_dataset_size, requires_samples_ordered_by_label)
