+++ modyn/tests/trainer_server/internal/trainer/remote_downsamplers/test_get_tensor_subset.py
@@ -1,6 +1,6 @@
 import pytest
 import torch
-from modyn.trainer_server.internal.trainer.remote_downsamplers.abstract_remote_downsample_strategy import (
+from modyn.trainer_server.internal.trainer.remote_downsamplers.abstract_remote_downsampling_strategy import (
     get_tensors_subset,
 )
