+++ modyn/tests/trainer_server/internal/trainer/remote_downsamplers/test_remote_gradnorm_downsample.py
@@ -1,8 +1,8 @@
 import torch
-from modyn.trainer_server.internal.trainer.remote_downsamplers.abstract_remote_downsample_strategy import (
+from modyn.trainer_server.internal.trainer.remote_downsamplers.abstract_remote_downsampling_strategy import (
     get_tensors_subset,
 )
-from modyn.trainer_server.internal.trainer.remote_downsamplers.remote_gradnorm_downsample import (
+from modyn.trainer_server.internal.trainer.remote_downsamplers.remote_gradnorm_downsampling import (
     RemoteGradNormDownsampling,
 )
 from torch import nn
