+++ modyn/trainer_server/internal/trainer/remote_downsamplers/__init__.py
@@ -1,7 +1,7 @@
 import os
 
-from .remote_gradnorm_downsample import RemoteGradNormDownsampling  # noqa: F401
-from .remote_loss_downsample import RemoteLossDownsampling  # noqa: F401
+from .remote_gradnorm_downsampling import RemoteGradNormDownsampling  # noqa: F401
+from .remote_loss_downsampling import RemoteLossDownsampling  # noqa: F401
 
 files = os.listdir(os.path.dirname(__file__))
 files.remove("__init__.py")
